Hide filters if #collapseFilters in URL
The filters in a page like https://network.satnogs.org/observations/?norad=33493 when viewed in a narrow display take a lot of space placing the list of observations below the fold. This patch allows to at least link  to a page that has this hidden by default. Use case for this is Mission Control Projections with Tiled Application Windows.